Fix some python plugin bugs, rearrange folder structure, add background fetcher
It is now possible to fetch meta data automatically in the background without user interaction.
This is a "best fit" search, as it always takes the first search result from the python plugins, this might lead to wrong data
but will work most of the time and saves a lot of time wen a lot of files need to be added.

The folder structure is chanegd a bit as in the future beside the commandline program a nepomu kservice and browser plugin parts will be added to this repo

Fixed a few minor python plugin bugs that prevent propper data fetching

Added 3 new commandline options to help the VideoExtractor to determine the search parameters for movies/tvshows based on teh file name a bit better