Replace nco:PersonContact with the more generic nco:Contact
This helps to avoid cluttering the database as PersonContact is mostly used to refer to people known to the user.
It is also not clear how to deal with 'artist names' of movie/music stars.

The downside, no property for family/given name or suffix exist for nco:Contact and only the fullname will be saved.
It is questionable if this additional data is used somewhere on the other hand.