Allow to search with a different plugin when no search result could be found
This is pach from Yuen Hoe Lim with some modifications to adapt to the latest changes

The first selected plugin is the one configured as favorite plugin in the KConfig file.
Afterwards all available plugins will be used until a some metadata could be found or no plugins are left.

CCMAIL: yuenhoe86@gmail.com