Remove old metadata created by the extractor before new data is added again
If the old data is not removed, properties with a cardinality > 1 will still have the old and possibly wrong metadata attached to the file
This patch ensures only the newly selected metadata will be used.

All metadata will be attached with the KComponentData("metadataextractor").
This should ensure adding/removing works even if the library is used outside of the extractor