Automatically overwrite banners in the kio:copy job if necessary
If the metadata extractor is run simultaniously it happens that the same banner is
downloaded severaltimes at once. When the file is then copied from the temp folder
to the destination an error message is thrown.

To handle this case, the file will now be overwrtten automatically.
This can be done safely, as it ill in fact be the same image anyway.