tvdb/tvdb_mail can crash the pthon interpreter when the tvdb_api.tvdb_episodenotfound exception is raised
The tvdb python plugin seems to cause the KROSS::Interpreter to crash if tvdb_api.tvdb_episodenotfound is raised.
This happens when one asks for a episode number tha tdoes not exist in the season.

Current solution is to disable this explicit error and handle it like any othe unknown exception.