Prefilter/sort online search results to allow better automatic selection
The returned search results from all plugins will be checked against
the actual search parameter. The Levenshtein distance is used to get
a measure of the similarity of the results.

Returned search results that are not close enough to the actual search
are skipped, othewise the list will be sorted by this distance.

This improves document and music search a lot and makes the automatic
fetcher without suer interaction) a lot less error prone and allow it
to skip obvious wrong results instead of adding it.