Allow each plugin to disable levenshtein distance check
If a plugin does not specify a distanceString that should be used
for the false positive search result check. No check is used and
all search results are valid ones.

Needed to cope with the tvdbmal plugin, that adds alias support and
transforms the user search strings into something else.